Rebalance the whole server against one timing model
The numbers across the 24 plugins were not on a shared scale, and in a few
places they made the game unfinishable. This rebuilds all of them from one
model, documented in BALANCE.md, and validates the result with an
end-to-end playthrough simulation that reads these config files and plays a
free account from the starter pet to the final zone.

The shape asked for: fast early, slowing sharply later, with a visible
power spike on every zone entry. Measured result for a free player - five
zones inside the first half hour, 1.2h to Emerald Depths, 3h to End
Barrens, 14.9h to Genesis Core, with the gap between unlocks growing about
40x across the ladder (5 minutes to 3.4 hours).

The spike is structural rather than tuned in: you walk into a zone carrying
the previous zone's pets and need ~4 seconds a cube, then buy that zone's
packs and settle at ~1.6 - a ~2.7x throughput jump, every zone, every time.

Things that were not tuning problems:

- Every zone past Frostpeak was unbuyable. Unlock costs demanded 32 Packed
  Ice, 32 Magma Blocks, 32 Ancient Debris and so on, and nothing in the
  game produces any of them.
- Diamond income was flat across the entire game - one diamond per cube
  kill in the last zone exactly as in the first - while ranks priced in
  diamonds climbed 1.15^n. Cube tiers now carry a diamond-value.
- Cube HP and pet damage were int. A dark-matter-fused secret pet in the
  last zone deals ~1.8e10, which silently overflowed. Widened to long.
- Every upgrade station sat in Meadow capped at level 10, so Coin Boost and
  Damage Boost were stuck at +10% for the whole game.
- The skill tree's pet-slot nodes granted one slot each rather than one per
  level, because value-formula returns the total at a level, not an
  increment.
- The Meadow world boss had 500,000 HP against a squad dealing ~70 a
  second, and paid more than the first nine zone unlocks combined.
- Block-tree tiers asked for 2,500,000 Stone in a zone players leave after
  about 120 of them.

Pack stations now sell ten at a time when you sneak. Zone packs are priced
so a stay funds ~30 of them in Meadow rising to ~200 later, which is what
makes the arrival spike close in a couple of minutes instead of the better
part of an hour; at one per click that would have been several hundred
clicks.

Donor ranks were measured, not designed around, and are unchanged: VIP
2.2x, Celestial 7.1x. Almost none of that comes from the +20/+80 pet slots
- after auto-fusion a squad's top few pets dominate the bag, so 8 slots to
88 is about +10% volley, most of it wasted on a cube you already one-shot.
It is the coin multiplier and the premium auto-switch cooldown.

+++ plugins/yield-zones/src/main/java/me/dontshare/yieldzones/cube/OreCube.java
@@ -17,7 +17,7 @@ public final class OreCube {
     private final CubeBonus bonus;
     private final int glowEntityId;
     private final UUID glowEntityUuid;
-    private int currentHp;
+    private long currentHp;
     // Whether the temporary white "you're looking at this one" outline is
     // currently applied - toggled as a glow flag directly on this cube's own
     // body entity (see OreCubeService#spawnHighlight/despawnHighlight), not
@@ -77,12 +77,12 @@ public final class OreCube {
         return glowEntityUuid;
     }
 
-    public int currentHp() {
+    public long currentHp() {
         return currentHp;
     }
 
     /** Applies damage, clamped at 0 - never goes negative. Returns true once this brings it to (or below) 0. */
-    public boolean damage(int amount) {
+    public boolean damage(long amount) {
         currentHp = Math.max(0, currentHp - amount);
         return currentHp <= 0;
     }
